refactor: move ClusterInfo utils
